Small code cleanups
Just small cleanups and refactorings here and there.

Especially, I realized that for some transformations it's much more convenient to do the whole thing inside a .bind environment, which lets us work as if we're back mostly manipulating T instead of Result<T>.  (And because .bind takes a function of type T => (T|Result<T>), you can return either a T or Result<T> depending on what's appropriate for that particular return statement.  If there are no messages, just return a T, you don't have to lift it.)