Making changes play nice with GSuite
Addressed a few issues that occurred when trying to push to GSuite:

(1) For some reason I can't quite track down, the custom typeguard that Result.err and Result.warn use doesn't work when compiling with browserify/tsify.  (We use this to localize T to the position of T when T has a position at all.  But not every T has a position!  T might even be void.)

The error seems to come down to using `is <interface>` as the predicate here:

    function isPositioned(p: any): p is Positioned {

even though this works in our usual environment.  Although that has a different tsconfig, I couldn't find anything that would cause this.

Anyway, this works instead:

    function isPositioned(p: any): p is {pos:CellPos} {

(2) The client that was running on the GSuite backend (Code.js) was still asking the interpreter to .runChecks(); now it doesn't.

(3) Content cells weren't highlighting correctly, because we had dropped the relevant messages in between their generation (in TstSequence.addContent() when the TstHeader is created) and their collection (in CreateTST).  One of the intermediate .addContent() methods was still returning void rather than ResultVoid.