Small improvements to error detection/recovery
(1) We caught the error for the following grammar, but didn't have a specific error message or appropriate recovery:

VERB:, text
     , foo
     ,
replace test:, from, to
             , f   , m

(Hint, "text" != "test".)  This doesn't parse because "test" is a reserved word.  Now we have a parse rule that handles this situation and errors.

(2) When op cells were parsed as ErrorOps or couldn't be parsed at all (which results in ErrorOps anyways), we used to default to replacing them with their .child.  But this is inappropriate in the case of replacements and unit tests, this replaces the actual grammar the programmer probably cares about with things that may not be grammars at all. (E.g. a replacement rule is meaningless on its own.)  But we also don't want to default to .sibling in every case, this would be a silly chose for a flubbed "table:" or assignment.

So the rule now is, default to .sibling, unless that's TstEmpty, in which case default to .child.

(3) I think I forgot to mention this in an earlier commit message today or yesterday, but it's now okay to name a sheet with a reserved word.  Yes, that sounds like a weird thing to allow, but it doesn't actually cause a problem, people do try to do it, and the error message was troublesome anyway (unlike most errors, it doesn't come from any particular cell).

For example, I've tried to name a sheet "Test".  This is a very reasonable sheet name, for experimental work or to hold your unit tests, but it wasn't allowed because it was a reserved word.

But why not?  Sure, you can't explicitly ASSIGN things to reserved words because those will get misinterpreted by the parser.  (I just realized today that the check we were doing for those never caught any, because the parser doesn't allow reserved words to be parsed into assignments in the first place.)  But if something is assigned to a reserved word for other reasons, there's no syntactic trouble REFERRING to it.  References are content, they don't get parsed as ops/headers anyway.  So for now at least, naming sheets "Test" and such is allowed again.