Checking for non-literal content in tests
Unit tests are not supposed to be arbitrary grammars, such that we're checking whether it's a subset of the main grammar.  They're meant to be just tables of literals.  (They *could* be arbitrary grammars, but I noticed in practice that when non-literal content was in one of these tables, it was a programming error, like just copy-pasting something from elsewhere and not realizing there was an embed in it.)

This commit introduces a transform that checks whether an op requires only literal content (currently only Test and TestNot) and removes non-literal headers from their child TstGrid.