Nomenclature changes
It was getting unwieldy to refer to everything as "Transforms", especially since their only important method was .transform.  In a typical micropass compiler like this, what we're calling a "Transform" is just a "Pass", so that's what we call them now.

Also, used more standard terminology to refer to the elements of a spreadsheet:

  "Sheet" -> "Worksheet"
  "SheetProject" -> "Workbook"

("Spreadsheet" is used in different ways by different people so I avoided it here.)