Refactoring of param system
(1) A bunch more renaming:

    -- The src/transforms directory is now src/passes (I thought I did this in the last commit but nope)
    -- The naming of grids/rows had become a bit unintuitive.  You'd think that a Grid and Row are pure syntactic objects without semantics, but actually this was PreGrid and PreRow.  So:
        -- TstPreGrid -> TstGrid
        -- TstPreRow -> TstRow
        -- TstGrid -> TstParamList
        -- TstRow -> TstParams

(1) The whole toParams/toParamList subsystem is eliminated.  Previously, when an enclosing op x required params, x.toGrammar() attempted to call child.toParamList() rather than child.toGrammar(), and got back a rather ugly result containing the necessary information to use those params.  This made the .toGrammar() pass an odd-one-out when it came to passes, since we were variably calling two different transform functions with different return types.  (Doing this sort of thing wouldn't be impossible in our new micropass framework but it'd be a case of... well, it'd be clever, and the whole thing is that we don't want to do all sorts of clever recursion anymore if we can avoid it.  So far as we can, we want our passes to be boringly recursive.)

But by the time we're creating ops now, we know a lot about the structure from all the checks and repairs we already did.  We can rely on param ops having their child be TstParamList, that having TstParam children, etc.  So rather than have a special function that recurses through an arbitrary tree (and fails when that tree has the wrong structure, which never happens anymore), we just have op.toGrammar() ask its TstParam grandchildren for the params it needs.

(3) Various nodes in the TST tree had structure that was no longer useful -- it was there from when they were created immediately during sheet parsing.  For example, every op needed to be a TstBinary so that the parser could give it a sibling during parsing, even if a sibling was semantically impossible for that particular op.  But now that these nodes are created relatively late, we've already parsed, validated, and (when necessary) gotten rid of semantically invalid siblings/children.  So most ops are just children of TstCellComponent again.

(4) Testing for non-literals in unit tests was coming too late to have any actual effect (beyond giving the programmer a warning); now it's before header/content association where it can actually make a difference.  This means that content cells underneath invalid unit test headers (e.g. "embed") now have the conventional warning that all other content cells under invalid headers have, so the unit tests have been updated to reflect this.