toGrammar() now a pass instead
This culminates a long string of refactorings, in which we slowly simplified Tst.toGrammar() until it met all the requirements to be composed with (before it) TstPasses and (after it) GrammarPasses.  Now the whole compilation pipeline (at least until we get into things like vocabulary and expression construction) has a consistent interface and can be composed.

The CreateGrammar pass is a bit of a beast now: because the input and output types are disjoint, it has to define a conversion for every input type, and a bunch of those are non-trivial.  As the interfaces of TSTs and Grammar are becoming more similar, though, this opens up the possibility for decomposing CreateGrammar into multiple, more-focused micropasses.