Converting GrammarPasses to use TST style
Passes for TSTs and Grammars were still using different double-dispatch mechanisms: TSTs were explicitly switching on the identity of the constructor, while Grammars were using the Visitor pattern.

The Visitor pattern had the nice property that it was automatic and type-checked, you couldn't put in a "transformsX" function and forget to hook it up to the dispatch code, get a type error somewhere, etc.  However, the ancestor class of visitor, the one that's mostly identity transformations, needs to know about every kind of possible object, and every object needs to know about it.  That's annoying in a language with weak and inconsistent module systems where recursive imports can cause weird errors.

The comparatively ad-hoc solution in TSTs is less safe; it relies a lot on explicit runtime checks/casts rather than automate that away using core language features.  However, it's very flexible and a pass doesn't have to know about anything beyond the kind of objects it operates on.

BEGIN SIDE DISCUSSION

Now that we have a ton of tree nodes turning into other kinds of tree nodes, that brings up the question of proving that, at each stage, only particular kinds of nodes are allowed to exist.  (E.g., at this stage only TstOp, TstNamespace, TstContent, and TstGrid are allowed to exist, and at the subsequent stage TstOp is out and a few more are in.)  This is the kind of proof that, in theory, we can use the type checker to prove.  (Zooming out, that's what a type checker IS, it's a proof assistant embedded in compilers.)

We were only making very broad claims about what things were allowed to exist at any point -- that before this pass, everything in the tree was a TstComponent, and afterward, also a TstComponent.  Before this pass, TstComponents, and afterwrd Grammars.  Before this one, Grammars, and afterward Grammars.  That lets us say things like "Okay, at the end, everything is Grammars and there are no TSTs left".

Can we be more specific?  T1 and T2 in Pass<T1,T2> can be anything, they can be union types of each specific class that's allowed to be in the input/output of that pass.

However, taking this through to its logical conclusion is a nightmare; it would end up needing us to also genericize the class hierarchy itself.  E.g., before Pass<T1,T2>, we would have to genericize the classes in the TST tree such that they don't just contain children of type TstComponent, but of type T1, and afterward of T2.  Otherwise mapping the micro-typed Pass to the children doesn't unify.

I also saw some instances where VSCode and the actual compiler disagreed about what can unify, which is unsettling.  Writing these kinds of "proofs" is hard enough without your two proof assistants disagreeing on what's grammatical.

Anyway, so we're not going to do that.

END SIDE DISCUSSION

Actually, we're going the opposite direction at the moment, towards there being an ultimate ancestor of all our tree objects.  TstComponent is the obvious choice here because there's almost nothing left in it except .mapChildren() and .msg().  So that's now Component, and Grammars are subclasses of it.

Side note: having rewritten the GrammarPasses, I am very suspicious about how they handle namespaces and embeddings... it seems to me that it shouldn't work, but it passes every test and "fixing" it breaks things.  Embeds have references to namespaces, so when the namespace changes, the reference in the embed has to change, too.  However, we're passing in the OLD namespace, and then end up returning a new one to which none of the embeds refer.  We should be able to show this doesn't work by taking a grammar that's changed by the final transformation, then referring to it by a symbol.  But I made that test and it works.

I'm going to leave it as-is for now, though, because I think the ultimate solution will make it moot.  EmbedGrammars really shoudn't be keeping references to anything anyway, they should just be strings and get their reference on-demand from an env.  The reason we don't currently do that is because it means passing an env into a TON of functions we don't currently -- calculateTapes, collectVocab, all of that stuff, and by experience I'm hesitant to touch that stuff while it's working.  Eventually, though, making this change will make things a lot cleaner, it's currently very hard to reason about references (as this discussion shows).