Several improvements/bugfixes to namespacing
This commit makes a bunch of small improvements/bugfixes to how we name symbols and handle namespaces of various sorts.  (I had gotten concerned about namespaces in the last commit, and that got me looking at some of this for the first time in a while, and I saw a bunch of things I'd overlooked.  But I still haven't addressed the concerns that I originally had.)

(1) We hadn't been checking (or doing anything to fix) namespaces in inappropriate places, like as arguments to operators or directly under another namespace.

(Actually, this latter is possible in theory, it could just be the default argument of the parent namespace... HOWEVER, it wouldn't serve any purpose, since the resulting qualified name would be the same as if the child symbols were in the parent namespace, and could cause reference ambiguities.  Better to just tell the programmers they're not allowed to do it at all.)

This commit adds a check for inappropriately positioned namespaces and a bunch of sheet tests to test it.

(2) Aside from the question of whether or not we allow it, there was a bug where a namespace child of a namespace (i.e. without an assignment in between) would cause anything declared before it not to exist.  This certainly isn't an appropriate response to this kind of programmer error!

This was because, during CreateNamespaces, we mapped children before transforming the parent, but then when mapping the parent iterate through its children expecting them to be TstEnclosure or TstEmpty.  That's not the case if we've already transformed the children, some of those can be TstNamespace already!  Making these functions late-mappers (who map their children as late as possible rather than early) fixes this.

(3) I was looking at getting more use from the Namespace<T> object, beyond just tapes, and found that it was causing some slowdown in deeply-nested rules.  This was because of an old change.

Originally, we used to handle renaming by wrapping a renaming adapter around it, but I was concerned that we could blow the stack in very long rule cascades.  So I changed this to a less-efficient but non-recursive version that just cloned the namespace and did the requested renaming.  (It's quadratic rather than linear, but when we were just doing like 5 nested rules, it doesn't really matter.  We had MUCH worse problems with deeply nested rules than a quadratic penalty.)

Anyway, now that nested rules are a lot faster, this penalty is becoming apparent.

So now we're back to doing it recursively. It's a bit different from before, rather than a renaming wrapping, namespaces form a stack, and upon failure a namespace forwards the request to the next one down. This is probably a little less efficient than the original for renaming itself, but is more flexible for other purposes.  (But our actual test grammars don't show a reliable slowdown compared to the original method.)

(4) How the QualifyNames pass handled its internal name/namespace stack was awkward (putting tuples into a stack) because we needed those for different purposes.  This was actually a remnant of when passes were genericized to args:T, all the arguments of a function had to be mushed into one.  It's clearer what's going on when they're separate arguments.

(5) We were assigning too many things to __DEFAULT__.  This happens when we create namespaces, and when I rewrote that code, I forgot why we actually assigned things to <namespace>.__DEFAULT__.  I thought this was the code that actually made default references work, but it isn't, that's actually later in QualifyNames.  All this was for is that we have to assign the relevant grammar to SOMETHING, so that it's there in the namespace's symbol map -- if it's not there, it's nowhere, and QualifyNames will never find it to do anything with it.

That is to say, you don't have to assign *every* last child of namespaces to __DEFAULT__, just the ones that have no other names.

(6) Also, I changed the string you assign it to to the empty string, so that the resulting qualified name is correct already.  That is, so we don't assign it to "Sheet3.VERBS.__DEFAULT__" and then assign it again to "Sheet3.VERBS", we just assign it to "Sheet3.VERBS" in the first place and then its re-assignment is skipped.

This also means that there isn't any name cleanup necessary in interfaces.  E.g., sidebar.html looked for symbols that ended in "__DEFAULT__", and didn't put them in the interface, in the knowledge that a simpler/less-ugly symbol referring to the exact same thing would be following later.  Now .allSymbols() only returns the simpler/less-ugly symbols in the first place, so interfaces don't have to know about them to hide them.

(7) Also, sidebar.html now hides the empty symbol "".  You can't see it, after all, and it's always duplicated by a symbol you can see (the one with the same name as the sheet you're on).  But unlike the __DEFAULT__s, .allSymbols() DOES return this one as a possible symbol.  (The CLI, after all, generates from that symbol by default.  So I think it's reasonable to put it in .allSymbols(), it really is a symbol that you can use.  It's just we don't want to display it in a user interface because it looks like an error of some sort.)