Fixed endless recursion of tapeclass detection
The recursion-protection of getTapeClass() was bugged, and I didn't notice because testRecursion had been turned off.

It was just a simple error, we were breaking recursion if an answer was cached, but only caching the answer after the recursive call, so it just recurses until the stack blows.

But really, we don't need the answer, <false, false> is enough right there.  (getTapeClass() in general has an OR semantics, so if it's true anywhere, it's true.  So if the previous call, the one that would have cached the real answer, has a true for one of those fields, the whole answer will be true for that field regardless of what this call returns.  So just return <false, false>.)

This also re-enables and updates testRecursion, which we should have been doing all along.