Fixed Expr.disjointBitsetDeriv and NegationExpr.bitsetDeriv to handle EpsilonTokens correctly.
Also fixed a bug in OutputTrieLeaf.toDict that only shows up when using bitsets.
The bug was introduced in fixed to toDict introduced when implementing EpsilonToken,
where we were incorrectly returning {} if we had no other results to return.

Finally, to aid debugging, we now show bits info in the id for EntangleExpr, and
we output bits and entanglement info for when logging in openDeriv when using
bitsets.