Reworking namespaces/embeds
A few commits ago, I got suspicious about our handling of namespaces inside of GrammarPasses.

GrammarPasses are (or were) special because what they're doing isn't just a tree-to-tree transformation.  Grammars really form a graph, because EmbedGrammars contain a reference to an NsGrammar, the one that contains the symbol to which EmbedGrammar.name refers. When you transform an NsGrammar, all the EmbedGrammars that refer to it have to change their reference to the new one.  That means passing around references to NsGrammars, too, which is why (at first) all GrammarPasses had a different API where you passed in the new NsGrammar, and later had a special env.ns in their env.

I got suspicious because, by my reading of our code, we were doing this wrong, with the result of EmbedGrammars referring to the wrong NsGrammars.  But our tests work, and trying to change the Ns/Embed logic made all the relevant tests fail.  So clearly I was missing something.

But my reading our code wrong is a big problem.  If Pat right now can't understand the logic of Pat six months ago, then that's some bad logic and should be replaced with something that will make sense to future Pats and Darlenes and others.

So this commit changes the whole logic.

(1) Now, EmbedGrammars do NOT contain references to any other grammars, they're just names.

(2) In any method where EmbedGrammar would be consulting its referent, we have to pass in an env parameter from which it can grab the current namespace.

(3) The env parameter no longer references NsGrammar as such, it just has a Namespace<Grammar>, env.symbolNS.  In recursive algorithms that take env, NsGrammar pushes its own symbols into env.symbolNS.

This adds a bunch of parameter mess onto a lot of functions, but in the long run that's probably not bad.  The last remaining part of the architecture that's really weird/hard are all those vocab/tape methods after the last grammar pass, but before constructExpr.  Getting their parameterization more similar to a Pass means we can look at refactoring those into actual Passes.