Changed testMatchOutput to use expect.deep.include.members to compare arrays.
Before we iterated over the outputs/expected_outputs calling expect.deep.include
to compare that result with the other array. Now we call expect once in each
direction having it compare the full arrays. This may be only a little faster,
but it's certainly cleaner code. It also appears to show all the differences for
the first failing expect, whereas before we just saw one because we compared the
elements individually.