Bugfixes for embedded replacements
Embedding replacements in a grammar (e.g., doing a morphophological rule on a subdomain like a stem, then assign that to STEM and using it elsewhere) was bugged in several ways.

(1) EmbedGrammar.mapChildren was just returning itself.  That seems fine, but it's not, because .tapes is memoized, and Passes can change the tape structure of the tree.  So you can't return `this.msg()` here, it has to be `new EmbedGrammar(this.name).msg()`.

(2) getVocabCopyEdges wasn't getting the vocab copy tapes right in certain circumstances.  (Specifically, in the circumstance arising from several symbols referring to the same grammar, which happens with defaults.)  Because of renaming, we can visit the same symbol with a different tape structure, but we exit early because we've visited that before.  This is the same problem that beset collectVocab, and so it needs the same, more nuanced early-exit strategy, where we cache both the symbol name and tape name.  This, in turn, means that getVocabCopyEdges needs to be parameterized by tape.

(3) In a convoluted bug that I'm not sure had any consequences, but is still good to squash, ReplaceGrammar.constructExpr() was calling .getAllTapes() to get the tapeNamespace it needs, but this calls .collectAllVocab().  Unlike most things that operate on namespaces, .collectAllVocab() doesn't fail on an unfound tape name, it just creates that tape.  (This is the function where tapes are constructed in the first place.)

So there's a problem any time we call .collectAllVocab() from anywhere but the root grammar.  From the vantage point of a non-root grammar, tapes may have different names, and calling .collectAllVocab() there will MAKE them exist under those names.  For a ReplaceGrammar inside a "replace text":, that's always the case.

So instead of creating a new TapeNamespace inside of ReplaceGrammar.constructExpr, we need to pass it in as a parameter.