Experiments with a new priority system
(1) When we found that cascades of replacement rules were getting slow, we changed how cascades were arranged so that they got the right tape priority.  This didn't work for every configuration of replacements, though, only for cascades expressed as a TstGrid, not (for example) "replace text:" operators one after another.

Backing up, there was a longstanding issue of calculating tape priority.  What we were doing was naive, but it worked well enough for long enough.  Now we've reached the point where our programmers are using patterns for which our old approach cannot get a reasonable answer.

We used to treat tapePriority as something that we could basically read off the order of the string[] results of calculateTapes().  Where we wanted a tape to be evaluated earlier (e.g. a tape that's involved in a join, so as to fail earlier), we adjusted the results of the relevant .calculateTapes() method so this was earlier in the string[].

When it turned out that the way in which we nested rule cascades was coming up with a bad order, we just changed the structure in which we nested them so that we got a better order.  But this really only worked for rule cascades expressed as blocks in the code.  Ultimately, there's no way that approach could get the right answer for a rule applied to an embed, then concatenated with other stuff, then wrapped in another rule. (But this is now what our programmers need!)

So this commit takes another approach: calculating for each tape a numerical estimation of its priority, then sorting the results to get the tapePriority.  The actual estimation process is currently also pretty silly, but it works for now.  We can potentially experiment with more sophisticated things later in the same basic framework.

(2) This commit also reincarnates the old RuleReplace transformation, now called ReplaceRulesOld, and updates it to the new Pass framework.  We had gotten rid of this because it resulted in a nesting order that gave bad results in the old priority system.  Now with the new priority system, it gives good results again.

(3) HideGrammar was erroneously missing some of the newer methods used for tape analysis and vocab collection, and HideGrammar and RenameGrammar were sometimes not properly short-circuiting when tapeName == toTape.  This was leading to some incorrect results in vocab collection, that only came to light when switching back to the old rule-nesting structure.

(4) There was, in theory, the ability to pre-name the hidden tapes of rules and hides (e.g. so that it refers to a unique location on the sheet).  But in practice we weren't managing to thread that information all the way from CreateGrammars to tape analysis.

Now we name rules according to their sheet/row, don't drop or mangle this name along the way, pass it on to the appropriate HideGrammar, don't drop or mangle this either.  As a result, we can now know which rule is responsible for different hidden tapes, which makes planning out a good tape priority easier.