Priority now uses vocab information
In this commit, I adjust the priority value for a tape according to its vocabulary: tapes with larger vocabularies now have higher priority.  The idea is that if tape T1 has a vocab of 2 and T2 has a vocab of 20, get to T2 first because it narrows things down more quickly.  This results in a small speedup (about 10%) on Michif.