Example of empty-to garden path
In testSheetsRules there's a new test that, while passing, illustrates the garden path that can happen when there's an empty-to rule in a rule cascade.  Doing this results in the generator consider paths through the output space that are never going to succeed, but that aren't actually determined to be invalid unil much later, leading to an exponential time penalty.  This is particularly bad when there are two such rules -- which isn't uncommon in a real grammar!

(Note: I switched to LTR to be able to reason about the outputs better, let's remember to change this back later.)