Grammar.collectVocab now returns string set
Grammar.collectVocab used to return void, and the vocab collection was a side effect -- we passed in a tapeNS argument, then registered vocabulary with that.

In this commit I began experimenting with the idea that tapeNSes and vocabs and such actually have a place in the grammar, as opposed to being a member of Interpreter.  (This is in service to a longer-term goal that every tree root have a denotation, which isn't true right now, but would simplify a lot of things if it were.)

This ends up hard-to-reason-about if vocabulary collection is done as a side effect, so this commit makes it purely functional, returning the set of possible tokens and not affecting anything else.

This also revealed a minor bug in UnitTestPass, we were getting tape priority before collecting the vocabulary, but conceptually tape priority requies knowing the size of vocabularies.  (In practice it doesn't matter much, all tests where the vocabulary size changes are going to fail anyway.)