Reworking of symbol namespaces for Exprs
In several recent commits we've been moving towards what I'll call the "denotational ideal", that every tree root object should have a denotation on its own.  Things like vocabs and symbol tables thus shouldn't be properties of (for example) Interpreter, because this means the denotation of any tree X is relativized to something outside of itself.  Instead, things like vocabs and symbol tables etc. should occupy nodes in the tree and push their changes onto an env parameter for children in their scope.

We started to do that for Grammar-related namespaces, but within the Expr hierarchy, EmbedExprs still had symbol tables that they were constructed with.  (I had actually forgotten about this, I think in a recent commit message I claimed otherwise.)

Changing this around was a little hairy, but it should start to simplify later as more things in this stage of compilation approach the denotational ideal.

(1) Interpreter used to hold a .symbolTable member, that was passed to constructExpr and which provided EmbedExprs with the symbol table they needed.  But this violates the denotational ideal, so it's gone now.

(2) Now instead there's a SymbolNsExpr, parallel to NsGrammar and created by it, that sits in the tree and, when a recursive algorithm like deriv or delta flows through it, it pushes its symbols onto the env.

(3) Before, when we generated from a non-default symbol, we did this by calling NsGrammar.getSymbol().  That's not going to cut it anymore, because NsGrammar.getSymbol isn't an NsGrammar, and thus doesn't create a SymbolNsExpr, and thus the symbol namespace ends up empty during delta/deriv.

So now NsGrammar has a member .selectedSymbol, and a method .selectSymbol(), that returns a new grammar except with the requested symbol as the .selectedSymbol.  When .constructExpr() is called, .selectedSymbol determines which expr is the child of the resulting SymbolNsExpr.  That gets us generating from the right expression and sets up the symbol namespace correctly for the resolution of embeds.

(4) Threading this into unit tests remains awkward, though.  Like before, UnitTestPass receives a symbol table in its constructor, and uses this when running unit tests.

I don't love this, though, it doesn't feel like it gels with the denotational ideal, but the solution in (3) isn't available here.  Unit tests are the one place we generate that's NOT generating from any symbol, not even a default symbol.  Maybe we eventually need to define a hidden symbol for the target of unit tests, so that we can generate from it using the same machinery as other generations.

(5) This refactoring revealed an unseen bug that's been with us for a while.  (A minor bug in that it's extremely unlikely to surface, but major in that it causes an exception to throw.) I think I had a vague notion this bug existed, but doing this snapped into place the test that would reveal it (now added to testEmbed).

The problem is that, at the end of generation, all exprs must be reduced to their most simplified form.  That's what constructX methods are for, they are intended to guarantee that only simplified exprs exist.  But this is difficult for EmbedExprs!  They don't necessarily know what they are yet, at the time of their construction, because they might refer to something that isn't constructed yet.

Now, most of the time this doesn't lead to bugs, because Expr.delta() and deriv() also create simplified forms.  As long as there's been a single delta() or deriv() call, we're good!  And there will be at least one such call if there's at least one tape.

So the only time this trouble surfaces is when there are no tapes.  The test that shows this is to create an embed that refers to something that will only be defined later, then make that an epsilon... and that's the whole grammar.  A bizarre grammar to write, for sure, but possible, and we don't want to be throwing exceptions because our invariances are violated.

There are a few ways to address this, but for the time being, a simple fix is to say that epsilons and nulls do define a tape, .END.  So there are no longer any grammars with no tapes at all.  This feels like cheating but it may actually help elsewhere, because the idea of "A grammar with no tapes" has always been a bit hard to think about.