Namespaces -> Collections
The original motivation for adding the "namespace:" op was to make it possible to express a multi-sheet project as an equivalent single sheet (for easy loading into apps).  To do this, we needed an op that had the same semantics as a sheet, but could be placed inside another sheet.

We didn't, at that time, anticipate what human programmers would want to use explicitly-defined namespaces for -- usually, grouping things together for easy access as a whole (e.g. Verb) but also with the ability to drill down further (e.g. Verb.Transitive).  The old semantics of namespaces mean, however, that using the namespace name on its own was a reference to whatever is defined last in that namespace (e.g. Verb refers to Verb.Intransitive or whatever happens to be defined last, not verbs in general).

The default behavior of namespaces has been troublesome before that, too. Sure, it's nice that (when you write so that the last thing in the sheet is the highest in the hierarchy) the highest thing is selected by default in the UI, but not everyone wants to write like that.  Some of our programmers want the highest-level thing at the *beginning* of a sheet, so it's the first thing you see when you read from the top, and it acts kind of like a table of contents for what follows.

So this commit makes some major changes to the interpretation of namespaces, so that all three interpretations (default first, default last, and default everything) can be achieved by some means.

(1) "Namespace:" is now called "collection:"

(2) If the programmer assigns to a symbol "Default" (case insensitive) under a collection op, that is now the default interpretation when the collection is referred to as a whole.

(3) If the programmer does NOT assign to symbol "Default", we automatically generate an alternation containing everything in the collection, and that is the default interpretation.

(4) These rules apply to worksheets as well, so it's important to think about which symbol (if any) is Default if you want to use default references to that worksheet.

(5) These rules DON'T apply to workbooks (that is, whole Gramble projects); the parser auto-sets the default symbol differently there.  While workbooks still become collections internally, they work like they did before, where the worksheet specified by the programmer becomes the default.  (I.e., the sheet they're looking at in the interface, or the sheet they specify as the source file on the command line.)  It'd be completely chaotic if, by default, generating from a project sampled/generated from every worksheet in the project.

Some implementation details:

(A) Everything that used to be called "Namespace" in relation to the specific Gramble operator/object is now called "Collection".  "Namespace" should now only refer to the data structure that's a stack of maps.

(B) We used to have an issue where, since some content could be referred to either by its programmer-assigned name or a default symbol, that grammar would get duplicated under its namespace during name qualification.  It would exist twice, and the two copies would not necessarily result in identical grammars after other passes applied.  (For example, any pass that increments a counter results in the two grammars becoming different.  This used to be the case for passes like rule replacement -- we would end up with too many tapes with different names, because each copy of that grammar would get a different set of new tapes.  We've since stopped naming those tapes using an index and so this wasn't an issue, but overall you can see why it's worrisome to duplicate grammars like this.)

The solution to this is actually easy.  Now when we make default symbols, they contain an *embed* of that grammar (or an alternation of embeds), rather than the grammar itself.

(C) Unnamed things under namespaces used to be valid or not -- if it was the last thing in a namespace, it was valid because it'd be assigned to default, and otherwise it was invalid and got a warning.

Now that the default interpretation of a collection (in the absence of an explicit "default:") is to refer to everything in it, all of these elements are potentially interpretable... but even so, having unassigned content under a namespace isn't a pattern we want to encourage.

What we do now is auto-assign unassigned content under a namespace to a special symbol _Auto<lineNum>, and issue a warning about it.  This symbol is handy to have around because it's easier to make an embed out of (when making the auto-default), and also it shows up in the sidebar UI as something you can query.

(D) In creating assignments during the CreateOps pass, we were grabbing the text of the cell from TstOp.text and trimming it to determine the name of the symbol.  We should have been using the TstOp.op.text instead, in case for some reason these differ -- it just happened that they never differed because of the ways/times we created symbols.  Now it happens that they differ sometimes (e.g. _Auto symbols systematically differ from their cell text), and TstOp.op.text is the one we want.