Regularizing test suite names
While we're busy changing nearly every test suite file, it's a good opportunity to regularize their names.  I was getting confused which things were plural vs not (Is it testJoins or testJoin?), confused when the name of the test file differed from the command (Is it testSheetsMulti or testMultiSheet?), and also finding cases where test suites weren't being called properly or hooked up to commands or commands referring to nonexistent files.

Now, the command is now always the name of the file (so testRegex.ts will always be called by "npm run testRegex").  Also, there are no more plurals, everything is singular (so testSheetsBasics -> testSheetBasic etc.)