Small cleanups and refactoring
(1) There were a few places where we needed a Set of tuples, but couldn't use them because tuples are compared by reference.  In the various vocab collection steps, we needed a Set<[string, string]> to prevent re-analyzing a symbol in the same tape context, but to do this we had to keep constructing a string key from the [string, string] and checking for that.  Meanwhile, vocabCopyEdges is conceptually a Set<[string,string]>, but we just used [string, string][] because nothing went wrong if edges weren't unique, we'd just re-copy.

This commit introduces what we really needed, ValueSet<T>, which is like Set<T> except it stringifies T (using JSON.stringify) to decide whether it's already in the set.  So you can use it with tuples, or anything where comparing JSON strings can be used as a proxy for equality.

I think avoiding the vocab re-copy speeds up vocab collection; I wasn't keeping careful track of how long it took for various test grammars, but my rough feeling is that it cuts the time in half.

(2) At some point, all grammars have to be wrapped in a CollectionGrammar (formerly NamespaceGrammar), which simplifies some of what follows (otherwise various steps have to consider whether or not they're operating on a bare grammar or a collection).  This was once done in the constructor method of Interpreter.  But that's not a great place to put it -- it happens that every grammar that isn't already a CollectionGrammar passes through that method, but that's now an inherent property of non-Collection grammars, that's just incidental.

Really, this should be guaranteed to happen sometime before the QualifyNames pass, which is the first place we need this.  So let's just put it before that as a pass.  (It doesn't need to be its own pass, though, I think it's reasonable to put this as part of the pass immediately before QualifyNames: AssignDefaults.  That's basically what it's doing anyway, it's assigning the input grammar to the default symbol of the output grammar.)

(3) We used to have PRE_GRAMMAR_PASSES and GRAMMAR_PASSES.  Now it's SHEET_PASSES (for the transformation from grids of cells to Grammars), NAME_PASSES (for the two transformations that deal with symbol names in particular), and GRAMMAR_PASSES.

Slowly it's becoming clear that there's a conceptual fourth stage after this, which is modifications to the grammar that result from specific queries to the grammar (user queries or unit tests), and operations whose results can change (like vocab collection) depending on the query.  These aren't Passes yet (my refactoring to do this broke too many things and I backed off), but we can foresee a future QUERY_PASSES stage as well.

(4) We had a naming issue where some things were Test/Testnot and others were UnitTest/NegativeUnitTest.  There was a rationale for this (basically a syntax/semantics distinction), but it's still confusing which one to use sometimes.  Shorter form won, it's now consistently Test/Testnot everywhere.

(5) I had removed Interpreter.getTapeNames because I thought nothing used it, but the CLI uses it, so now it's back.